blutit: Introduce new detect_episodes() function
Eventually this will be able to detect a set of TV series episodes;

By calculating the average duration and then the standard deviation
from that mean, we can identify all titles that have a duration of
that average +/- the standard deviation. All that fall into this category
are likely to be episodes from the same TV series as typically when aired
they are the same length.