all: Augment music hierarchy with genre
This patch effectively reverts back to having the genre in the hierarchy.
This was originally removed when we switched to Musicbrainz because for
some reason Musicbrainz/ABCDE failed to populate that variable (GENRE)
with any value. We now, for the time being at least, switch back to using
CDDB.

All scripts and config files have been modified to support this additional
depth.